feat(about-me): add new components and refactor about section
- Create new reusable components (CardWrapper, StoryCard, TechStack, Achivements, AboutTitle, BackgroundOrbs)
- Extract constants to separate file
- Refactor AboutMeSection to use new components
- Update hero content with education details
- Add new CSS animations and gradient text styles
