+++ components/about-me/constants.ts
@@ -1,0 +1,32 @@
+import { easeInOut } from "framer-motion";
+
+export const techStack = [
+  "React",
+  "Next.js",
+  "TypeScript",
+  "Tailwind CSS",
+  "Node.js",
+  "Express.js",
+  "MongoDB",
+  "Firebase",
+  "Git",
+];
+
+export const achievements = [
+  "Completed a 90-hour React and Next.js training with Gaza Sky Geeks, where my team 'NextCodeGen' secured first place in the final project phase with our app 'Codepilot'.",
+  "Continued to deepen my skills by building several personal projects using pure React and a large-scale application with Next.js.",
+  "Explored JavaScript in depth, gaining a strong understanding of how the language works under the hood.",
+  "Applied and was accepted into the Gaza Sky Geeks Backend Code to Career Bootcamp after passing multiple screening stages and a technical interview."
+];
+
+export const myStory = `
+    After the war in Gaza and the closure of universities, I took the opportunity to begin learning the fundamentals of web development — starting with HTML, CSS, and JavaScript. I built several projects using vanilla JavaScript, which helped me understand the core of how the web works.
+    As I looked for the next step, I discovered a training opportunity with Gaza Sky Geeks focused on React and Next.js. I applied, passed the interview, and was accepted. During this intensive program, my team and I built a project called Codepilot, which won first place in the final demo phase.
+    After graduating, I focused on deepening my knowledge of JavaScript and building more projects using both React and Next.js. Driven to grow as a full-stack developer, I applied to the Gaza Sky Geeks Backend Bootcamp, where I was accepted after a competitive multi-stage selection process.
+    Now, I’m continuing to expand my skill set by learning React Native while exploring backend technologies — on a mission to become a well-rounded software engineer.
+`
+
+export const floatingAnimation = {
+    y: [-10, 10, -10],
+    transition: { duration: 6, repeat: Number.POSITIVE_INFINITY, ease: easeInOut },
+  }
